+++ frontend/src/views/ensemble/SingleNote.tsx
@@ -3,13 +3,15 @@ import { FC, useEffect, useState } from "react";
 import { Box } from "@mui/material";
 import { useRecoilValue } from "recoil";
 import { beatNumberAtom } from "../../recoil/beat";
-import { Synth } from "tone";
+// import { Synth } from "tone";
+import * as Tone from "tone";
+import FluteDb5 from "frontend/src/assets/samples/FluteDb5.wav";
 
 // Colors for notes
 const colorMapping = {
   [NoteType.REST]: "lightgray",
   [NoteType.ATTACK]: "gray",
-  [NoteType.SUSTAIN]: "darkgray", // ???? why is this lighter than gray
+  [NoteType.SUSTAIN]: "darkgray", // This is lighter than gray for some reason
 };
 
 interface SingleNoteProps {
@@ -18,7 +20,14 @@ interface SingleNoteProps {
   pitch: number;
 }
 
+const now = Tone.now();
 const PITCH_VALUES = ["C5", "B4", "A4", "G4", "F4", "E4", "D4", "C4"];
+const fluteSampler = new Tone.Sampler({
+  urls: {
+    "Db5": FluteDb5,
+  },
+  release: 1,
+}).toDestination();
 
 const SingleNote: FC<SingleNoteProps> = ({ beatNumber, pitch }) => {
   // note type - attack sustain rest
@@ -29,11 +38,14 @@ const SingleNote: FC<SingleNoteProps> = ({ beatNumber, pitch }) => {
   useEffect(() => {
     if (!playNow) return;
     // Trigger a music note if we are not a NoteType.REST
-    // console.log("Playing Note")
-    // Insert Josh ...
-    if (currentNoteType !== NoteType.REST) {
-      const synth = new Synth().toDestination();
-      synth.triggerAttackRelease(PITCH_VALUES[pitch], "8n");
+    if (currentNoteType === NoteType.ATTACK) {
+      // const synth = new Synth().toDestination();
+      // synth.triggerAttackRelease(PITCH_VALUES[pitch], "8n");
+      // fluteSampler.triggerAttackRelease(PITCH_VALUES[pitch], "16n");
+      fluteSampler.triggerAttack(PITCH_VALUES[pitch]);
+    }
+    else if (currentNoteType === NoteType.REST) {
+      fluteSampler.triggerRelease(now);
     }
   }, [playNow, pitch, currentNoteType]);
 
@@ -60,9 +72,8 @@ const SingleNote: FC<SingleNoteProps> = ({ beatNumber, pitch }) => {
   // If hovering, overwrite the note color
 
   // Color depends on Hover and NoteType
-
   let noteColor = colorMapping[NoteType.REST];
-  let highlight = 1;
+  let highlight = 1;  // 100% Brightness which is Normal Brightness
 
   // Coloring
   if (currentNoteType === NoteType.ATTACK)
